refactor:将骰娘相关的代码单独打包到dicer目录下
